requirements.txt
openpyxl==3.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-telegram-bot==20.3
